Rename files in the pages folder and add images
