+++ .eslintrc.json
@@ -22,6 +22,7 @@
     "react/jsx-filename-extension": ["warn", { "extensions": [".js", ".jsx"] }],
     "react/react-in-jsx-scope": "off",
     "import/no-unresolved": "off",
+    "import/no-extraneous-dependencies": ["error", {"devDependencies": true}],
     "no-shadow": "off",
     "react/prop-types": "off",
     "react/prefer-stateless-function": "off",
