+++ src/app.js
@@ -1,17 +1,24 @@
 import React from 'react';
 import './app.css';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Calculator from './components/Calculator';
-import Navbar from './pages/Navbar';
+import Home from './pages/home';
+import Quote from './pages/about';
+import NavBar from './pages/navbar';
 
-class App extends React.PureComponent {
-  render() {
-    return (
-      <div className="App">
-        <Navbar />
-        <Calculator />
-      </div>
-    );
-  }
-}
+const App = () => (
+  <div className="App">
+    <>
+      <Router>
+        <NavBar />
+        <Routes>
+          <Route index element={<Home />} />
+          <Route path="/calculator" element={<Calculator />} />
+          <Route path="/about" element={<Quote />} />
+        </Routes>
+      </Router>
+    </>
+  </div>
+);
 
 export default App;
