+++ src/logic/calculate.js
@@ -1,4 +1,5 @@
-import operate from './operate';
+import { defaults } from 'autoprefixer';
+import Operate from './operate';
 
 function isNumber(item) {
   return !!item.match(/[0-9]+/);
@@ -13,7 +14,7 @@ function isNumber(item) {
  *   next:String       the next number to be operated on with the total
  *   operation:String  +, -, etc.
  */
-export default function calculate(obj, buttonName) {
+const Calculate = (obj, buttonName) => {
   if (buttonName === 'AC') {
     return {
       total: null,
@@ -68,7 +69,7 @@ export default function calculate(obj, buttonName) {
   if (buttonName === '=') {
     if (obj.next && obj.operation) {
       return {
-        total: operate(obj.total, obj.next, obj.operation),
+        total: Operate(obj.total, obj.next, obj.operation),
         next: null,
         operation: null,
       };
@@ -111,7 +112,7 @@ export default function calculate(obj, buttonName) {
     }
 
     return {
-      total: operate(obj.total, obj.next, obj.operation),
+      total: Operate(obj.total, obj.next, obj.operation),
       next: null,
       operation: buttonName,
     };
@@ -130,4 +131,6 @@ export default function calculate(obj, buttonName) {
     next: null,
     operation: buttonName,
   };
-}
+};
+
+export default Calculate;
