+++ src/logic/operate.js
@@ -1,6 +1,6 @@
 import Big from 'big.js';
 
-export default function operate(numberOne, numberTwo, operation) {
+const Operate = (numberOne, numberTwo, operation) => {
   const one = Big(numberOne);
   const two = Big(numberTwo);
   if (operation === '+') {
@@ -27,4 +27,6 @@ export default function operate(numberOne, numberTwo, operation) {
     }
   }
   throw Error(`Unknown operation '${operation}'`);
-}
+};
+
+export default Operate;
