+++ src/pages/home.js
@@ -1,13 +1,13 @@
 import React from 'react';
+import math from '../images/math.png'
 
 const Home = () => (
-  <section className="w-full pt-20 relative h-screen bg-gradient-to-r from-gray-800 to-gray-900 text-white">
+  <section className="w-full pt-20 relative h-screen">
     <h1 className="text-4xl mt-28 mx-10 text-red-400">Welcome to Our Page!</h1>
-    {/* <img
-      src={landing}
-      alt="home-minimalist-bg"
-      className="w-full h-screen top-0 object-cover absolute mix-blend-overlay"
-    /> */}
+    <img
+      src={math}
+      alt="math-magic"
+    />
     <p className="text-base my-8 mx-10">
       Lorem ipsum dolor sit amet consectetur adipisicing elit. Eaque rerum atque
       optio quaerat consequuntur a facere quasi modi neque maiores veniam et
